Move CurrencyView to kernel
