+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/port/input/ContributorFacadePort.java
@@ -1,12 +1,12 @@
 package onlydust.com.marketplace.project.domain.port.input;
 
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 import onlydust.com.marketplace.kernel.pagination.Page;
 import onlydust.com.marketplace.kernel.pagination.SortDirection;
 import onlydust.com.marketplace.project.domain.model.Contributor;
 import onlydust.com.marketplace.project.domain.model.GithubRepo;
 import onlydust.com.marketplace.project.domain.model.Project;
 import onlydust.com.marketplace.project.domain.view.ContributionView;
-import onlydust.com.marketplace.project.domain.view.CurrencyView;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.List;
