+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/port/output/AccountingServicePort.java
@@ -1,6 +1,6 @@
 package onlydust.com.marketplace.project.domain.port.output;
 
-import onlydust.com.marketplace.project.domain.view.CurrencyView;
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 
 import java.math.BigDecimal;
 import java.util.UUID;
