+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/Money.java
@@ -3,6 +3,7 @@ package onlydust.com.marketplace.project.domain.view;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 
 import java.math.BigDecimal;
