+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/ProjectRewardView.java
@@ -2,6 +2,7 @@ package onlydust.com.marketplace.project.domain.view;
 
 import lombok.Builder;
 import lombok.Data;
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 import onlydust.com.marketplace.kernel.model.RewardStatus;
 
 import java.math.BigDecimal;
