+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/view/TotalEarnedPerCurrency.java
@@ -2,6 +2,7 @@ package onlydust.com.marketplace.project.domain.view;
 
 import lombok.Builder;
 import lombok.Data;
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 
 import java.math.BigDecimal;
