+++ project-domain/src/test/java/onlydust/com/marketplace/project/domain/service/RewardServiceTest.java
@@ -2,12 +2,12 @@ package onlydust.com.marketplace.project.domain.service;
 
 import com.github.javafaker.Faker;
 import onlydust.com.marketplace.kernel.exception.OnlyDustException;
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 import onlydust.com.marketplace.kernel.port.output.IndexerPort;
 import onlydust.com.marketplace.project.domain.model.RequestRewardCommand;
 import onlydust.com.marketplace.project.domain.model.Reward;
 import onlydust.com.marketplace.project.domain.port.output.AccountingServicePort;
 import onlydust.com.marketplace.project.domain.port.output.RewardStoragePort;
-import onlydust.com.marketplace.project.domain.view.CurrencyView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
