+++ project-domain/src/test/java/onlydust/com/marketplace/project/domain/view/ProjectBudgetsViewTest.java
@@ -1,5 +1,6 @@
 package onlydust.com.marketplace.project.domain.view;
 
+import onlydust.com.marketplace.kernel.model.CurrencyView;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
